+++ src/main/java/insane96mcp/iguanatweaksreborn/module/combat/AbsorptionArmor.java
@@ -35,9 +35,9 @@ public class AbsorptionArmor extends Feature {
             }
             if (attribute.equals(Attributes.ARMOR_TOUGHNESS)) {
                 if (modifier.getOperation() == AttributeModifier.Operation.ADDITION)
-                    toAdd.put(RegeneratingAbsorption.REGEN_ATTRIBUTE.get(), new AttributeModifier(modifier.getId(), modifier.getName(), modifier.getAmount() * 0.1d, AttributeModifier.Operation.MULTIPLY_BASE));
+                    toAdd.put(RegeneratingAbsorption.SPEED_ATTRIBUTE.get(), new AttributeModifier(modifier.getId(), modifier.getName(), modifier.getAmount() * 0.1d, AttributeModifier.Operation.MULTIPLY_BASE));
                 else
-                    toAdd.put(RegeneratingAbsorption.REGEN_ATTRIBUTE.get(), new AttributeModifier(modifier.getId(), modifier.getName(), modifier.getAmount(), modifier.getOperation()));
+                    toAdd.put(RegeneratingAbsorption.SPEED_ATTRIBUTE.get(), new AttributeModifier(modifier.getId(), modifier.getName(), modifier.getAmount(), modifier.getOperation()));
                 toRemove.put(attribute, modifier);
             }
         });